requirements.txt
kafka-python==2.0.2
pyspark==3.4.1
flask==2.3.2
pandas==2.1.2
